Moved subscriptions endpoints to channels module
